requirements.txt
openai==0.28.1
requests==2.31.0
websocket-client==1.6.4
sounddevice==0.4.6
numpy==1.24.3
pyaudio==0.2.13
python-dotenv==1.0.0
pygame>=2.0.0
PyOpenGL>=3.1.6
PyWavefront>=1.3.3
numpy>=1.21.0 